requirements.txt
# Core application dependencies
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
gspread>=5.10.0
google-auth>=2.22.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Development dependencies (optional)
# Uncomment for development:
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
